RunningLog | 받은 스탬프 ViewHolder 클래스 분리
