+++ app/src/main/java/com/applemango/runnerbe/presentation/screen/fragment/mypage/runninglog/detail/GotStampAdapter.kt
@@ -4,31 +4,14 @@ import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
-import androidx.recyclerview.widget.RecyclerView
-import com.applemango.runnerbe.R
 import com.applemango.runnerbe.databinding.ItemGotStampBinding
-import com.applemango.runnerbe.presentation.screen.dialog.stamp.getStampItemByCode
-import com.bumptech.glide.Glide
 
-class GotStampAdapter : ListAdapter<MemberStampData, GotStampAdapter.GotStampViewHolder>(
+class GotStampAdapter : ListAdapter<MemberStampData, GotStampViewHolder>(
     GOT_STAMP_DIFF_UTIL
 ) {
     private var listener : GotStampClickListener? = null
 
-    class GotStampViewHolder(
-        private val binding: ItemGotStampBinding
-    ): RecyclerView.ViewHolder(binding.root) {
-        fun bind(item: MemberStampData, listener : GotStampClickListener?) {
-            binding.item = item
-            binding.listener = listener
-            with(binding) {
-                Glide.with(itemView.context)
-                    .load(getStampItemByCode(item.stampCode)?.image)
-                    .error(R.drawable.ic_user_default)
-                    .into(ivStamp)
-            }
-        }
-    }
+
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): GotStampViewHolder {
         val inflater = LayoutInflater.from(parent.context)
