+++ app/src/main/java/com/applemango/runnerbe/presentation/screen/fragment/mypage/runninglog/groupprofile/ProfileAdapter.kt
@@ -1,33 +1,31 @@
 package com.applemango.runnerbe.presentation.screen.fragment.mypage.runninglog.groupprofile
 
 import android.view.LayoutInflater
-import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
-import androidx.recyclerview.widget.RecyclerView
-import com.applemango.runnerbe.R
 import com.applemango.runnerbe.data.network.response.JoinedRunnerResult
 import com.applemango.runnerbe.databinding.ItemGroupProfileBinding
-import com.applemango.runnerbe.presentation.screen.dialog.stamp.getStampItemByCode
 import com.applemango.runnerbe.presentation.screen.fragment.mypage.runninglog.otheruser.OtherUserProfileClickListener
-import java.lang.ref.WeakReference
 
-class ProfileAdapter: ListAdapter<JoinedRunnerResult, ProfileAdapter.ProfileViewHolder>(
+class ProfileAdapter: ListAdapter<JoinedRunnerResult, ProfileViewHolder>(
     profileDiffUtil
 ) {
-    private var otherUserProfileClickListener: WeakReference<OtherUserProfileClickListener>? = null
+    private var otherUserProfileClickListener: OtherUserProfileClickListener? = null
     private var selectedPosition = -1
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ProfileViewHolder {
         val inflater = LayoutInflater.from(parent.context)
-        return ProfileViewHolder(ItemGroupProfileBinding.inflate(inflater, parent, false))
+        return ProfileViewHolder(
+            ItemGroupProfileBinding.inflate(inflater, parent, false),
+            updateSelectedPosition = { position -> updateSelectedPosition(position) }
+        )
     }
 
     override fun onBindViewHolder(holder: ProfileViewHolder, position: Int) {
         val item = getItem(position)
         if (item != null) {
-            holder.bind(item, otherUserProfileClickListener?.get())
+            holder.bind(item, selectedPosition, otherUserProfileClickListener)
         }
     }
 
@@ -40,7 +38,7 @@ class ProfileAdapter: ListAdapter<JoinedRunnerResult, ProfileAdapter.ProfileView
     }
 
     fun setOnProfileClickListener(listener: OtherUserProfileClickListener) {
-        this.otherUserProfileClickListener = WeakReference(listener)
+        this.otherUserProfileClickListener = listener
     }
 
     fun updateProfileStampByPosition(position: Int, newStampCode: String) {
@@ -48,28 +46,6 @@ class ProfileAdapter: ListAdapter<JoinedRunnerResult, ProfileAdapter.ProfileView
             this[position] = this[position].copy(stampCode = newStampCode)
         }
         submitList(updatedList)
-    }
-
-    inner class ProfileViewHolder(
-        private val binding: ItemGroupProfileBinding
-    ) : RecyclerView.ViewHolder(binding.root){
-        fun bind(item: JoinedRunnerResult, otherUserProfileClickListener: OtherUserProfileClickListener?) {
-            val listener = View.OnClickListener {
-                otherUserProfileClickListener?.onProfileClicked(bindingAdapterPosition, item.userId, getStampItemByCode(item.stampCode))
-                updateSelectedPosition(bindingAdapterPosition)
-            }
-
-            if (bindingAdapterPosition == selectedPosition) {
-                binding.flProfile.setBackgroundResource(R.drawable.bg_g5_circle_shape_primary_stroke)
-            } else {
-                binding.flProfile.setBackgroundResource(R.drawable.bg_g7_circle_shape_no_stroke)
-            }
-
-            binding.item = item
-            binding.stamp = getStampItemByCode(item.stampCode)
-            binding.listener = listener
-            binding.clickListener = otherUserProfileClickListener
-        }
     }
 
     companion object {
